Refactoring LTC Python scripts: - The original ltc.py will now contain just API calls - Each script gets its own module with all associated code - A few more common functions extracted into utils.py
