+++ .gitignore
@@ -135,3 +135,6 @@ logs/
 *.swp
 
 .idea/
+
+.DS_Store
+outputs/*.csv
