+++ flask_server.py
@@ -4,7 +4,7 @@ import click
 
 from app import create_app
 from decouple import config
-from app.api import ltc
+from app.api import ltc, aggregate_outbreaks, close_outbreaks, data_quality_checks
 
 import config as configs
 
@@ -34,24 +34,24 @@ def deploy():
 @click.option('-o', '--outfile')
 @click.argument("url")
 def cli_aggregate_outbreaks(outfile, url):
-    ltc.cli_aggregate_outbreaks(outfile, url)
+    aggregate_outbreaks.cli_aggregate_outbreaks(outfile, url)
 
 
 @app.cli.command("close_outbreaks")
 @click.option('-o', '--outfile')
 @click.argument("url")
 def cli_close_outbreaks(outfile, url):
-    ltc.cli_close_outbreaks_nm_ar(outfile, url)
+    close_outbreaks.cli_close_outbreaks_nm_ar(outfile, url)
 
 
 @app.cli.command("quality_checks")
 @click.option('-o', '--outfile')
 @click.argument("url")
 def cli_close_outbreaks(outfile, url):
-    ltc.cli_quality_checks(outfile, url)
+    data_quality_checks.cli_quality_checks(outfile, url)
 
 
 @app.cli.command("check_data_types")
 @click.argument("url")
 def cli_check_data_types(url):
-    ltc.cli_check_data_types(url)
+    data_quality_checks.cli_check_data_types(url)
